refactor: move test files and update import paths to align with new directory structure
